move the actual `RedisSub` object into its own module
